+++ src/main/java/nextstep/courses/domain/model/Course.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.model;
 
 import org.springframework.lang.NonNull;
