+++ src/main/java/nextstep/courses/domain/model/SessionPeriod.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.model;
 
 import java.time.LocalDateTime;
