+++ src/main/java/nextstep/courses/domain/model/SessionStatus.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.model;
 
 public enum SessionStatus {
     READY, OPEN, CLOSED
