+++ src/main/java/nextstep/courses/domain/repository/CourseRepository.java
@@ -1,4 +1,6 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.repository;
+
+import nextstep.courses.domain.model.Course;
 
 public interface CourseRepository {
     int save(Course course);
