+++ src/test/java/nextstep/courses/domain/SessionImageTest.java
@@ -1,5 +1,6 @@
 package nextstep.courses.domain;
 
+import nextstep.courses.domain.model.SessionImage;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
